refactor #step2 - 패키지 분리(domain, utils, view)
