+++ src/main/java/step2/utils/RandomNumberGenerator.java
@@ -1,9 +1,9 @@
-package step2;
+package step2.utils;
 
 import java.util.Random;
 
 public class RandomNumberGenerator {
-    final static int LIMIT_OF_RANDOM_NUMBER = 10;
+    private final static int LIMIT_OF_RANDOM_NUMBER = 10;
 
     public int getRandomNumber() {
         Random random = new Random();
