+++ src/test/java/step2/CarTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step2.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
